jose-jwk: flatten `rcrypto` submodule
As there's currently only one crypto provider, there's no need for a
separate submodule.
