+++ jose-jwk/src/crypto/rsa.rs
@@ -10,8 +10,8 @@ use rsa::{
 
 use jose_jwa::{Algorithm, Algorithm::Signing, Signing::*};
 
-use super::super::KeyInfo;
 use super::Error;
+use super::KeyInfo;
 use crate::{Rsa, RsaPrivate};
 
 impl KeyInfo for RsaPublicKey {
